[fix] 07/15 검색 페이지 훅 분리 #111
